mb/google/nissa/var/glassway: Removed the flag of DB_1A for pmc_mux
Due to only 1A on DB, the test item of FAFT EC
firmware.DevInsertUSBScreen.insert_usb would be captured much error
message to detect C1 if set the flag.
And then missed to capture the correct message when running the test.
Therefore, removed the DB_1A for pmc_mux on daughter board to fix
the issue.

BUG=b:451436640
TEST=USE="${USE} -project_all project_craaskyu2" emerge-nirva \
     coreboot chromeos-zephyr chromeos-bootimage
     Confirm firmware.DevInsertUSBScreen.insert_usb PASS.

Change-Id: I6b1a3c99d422c99103818556365b0e5929a18dbf
Signed-off-by: Daniel Peng <Daniel_Peng@pegatron.corp-partner.google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/89538
Reviewed-by: Eric Lai <ericllai@google.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Kapil Porwal <kapilporwal@google.com>