mb/google/bluey: Condition slow charging enablement on charger presence
Update the logic to properly condition the slow battery charging feature.

Previously, enable_slow_battery_charging() was called solely if the
system was in a low-power boot state (is_low_power_boot()).

This commit adds an explicit check for
`google_chromeec_is_charger_present()` to ensure that the slow charging
feature is only enabled when a charger is physically connected.

TEST=Able to build and boot google/quenbi.

Change-Id: I24b6626343a25a4fab3f5d77c1d114e797781be7
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/90335
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Kapil Porwal <kapilporwal@google.com>